tplink-hs: Use a timer as the main session source
The timer stays registered all the time and is there to initiate polling
requests.

When a request is sent, a socket is created and registered as an event
source to get notified when the reply comes in. As the device closes
the socket after the reply, we also close the socket on our end and remove
it from the event sources.