appa-dmm: MEM/LOG support for legacy APPA 505 devices
Older models of the APPA 500 series using the legacy protocol have a
different method of obtaining MEM and LOG data.

This commit adds support for MEM and LOG download for these devices
and corrects endieness quirks related to them. As before, information
is mapped to the new protocol for compatibility.

Functionality tested with a Voltcraft VC-950.

Implemented in a way so that in future support for APPA 100 and 300
series legacy models can be added.