output/csv: samplerate and sample_interval from meta-data
For DMMs - in many cases, samplerate and sample_interval are not available
as a hardware config value but is provided using a meta frame.

These meta frames are now decoded and used as an optional source for the
samplerate and sample_interval. Some further testing concerding the
csv export functionality of pulseview might be needed.