appa-dmm: Support for APPA 300 and APPA 207 devices (EXPERIMENTAL)
Implemented the communication protocol of the APPA 300 series (also used
for APPA 207 desktop multimeter), covering APPA 301, 303 and 305.

Allows showing all major readings (Main and right display), units and
function.

Left reading is not parsed because it would mean a significant downside
for the other protocols. If someone comes around and needs it, it could
be done easily (extra channel needed).

The driver is flagged as EXPERIMENTAL and puts out a big fat warning
message if activated. It should work and it ran against a dummy
communication partner, but it is otherwise completely untested.