appa-dmm: APPA Series S, 170 / Sefram 72xx / BENNING MM 10, CM 9-2/10 / RS PRO Series S - minor fixes, now tested & working
RS PRO S2 was tested and a couple of minor datalog / mem issues have been
discovered and fixed. Type of memory (LOG or MEM) is now autodiscovered.
Detection of brand name still not implemented, but that is rather irelevant.

Series 170 can be considered to work as well since its communication and
memory handling is 100% identical to the Series S, it only differs in it's
hardware capabilities (clamp meter vs. handheld meter and features).

BLE Scan Names for those devices completed.