gwinstek-gds800: support current libserialport
This fixes a rare interaction between GDS-2000,
scpi-serial, and the gwinstek-gds800 driver.
Sometimes, due to the contents of the oscilloscope
data buffers, the oscilloscope would emit a newline
character in the middle of the binary datastream.

Very old (sometime before 0.1.0) libserial opened
used the serial device in canonical mode, which would
detect this as a partial transmission and return early.
Some support code in the gds800 driver would handle it.

But since 0.1.1 libserialport disables canonical
mode which makes this the gds800 driver's additional
processing unnecessary and harmful.

When the GDS-800 driver detected this rare condition,
it would cause the oscilloscope to read 0 volts for
whichever channel it occurred on.